Safari-611.1.10/LayoutTests/fast/url/idna2003-expected.txt

IDNA2003 handling in domain name labels.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The PASS/FAIL results of this test are set to the behavior in IDNA2003.
FAIL canonicalize('http://faß.de/') should be http://fass.de/. Was http://xn--fa-hia.de/.
FAIL canonicalize('http://βόλος.com/') should be http://xn--nxasmq6b.com/. Was http://xn--nxasmm1c.com/.
FAIL canonicalize('http://ශ්‍රී.com/') should be http://xn--10cl1a0b.com/. Was http://xn--10cl1a0b660p.com/.
FAIL canonicalize('http://نامه‌ای.com/') should be http://xn--mgba3gch31f.com/. Was http://xn--mgba3gch31f060k.com/.
PASS canonicalize('http://www.looĸout.net/') is 'http://www.xn--looout-5bb.net/'
PASS canonicalize('http://ᗯᗯᗯ.lookout.net/') is 'http://xn--1qeaa.lookout.net/'
PASS canonicalize('http://www.lookout.сом/') is 'http://www.lookout.xn--l1adi/'
FAIL canonicalize('http://www.lookout.net：80/') should be http://www.lookout.net:80/. Was http://www.lookout.net：80/.
FAIL canonicalize('http://www‥lookout.net/') should be http://www..lookout.net/. Was http://www‥lookout.net/.
PASS canonicalize('http://www.lookout‧net/') is 'http://www.xn--lookoutnet-406e/'
PASS canonicalize('http://www.looĸout.net/') is 'http://www.xn--looout-5bb.net/'
FAIL canonicalize('http://www.lookout.net⩴80/') should be http://www.lookout.net::%3D80/. Was http://www.lookout.net⩴80/.
FAIL canonicalize('http://www .lookout.net/') should be http://www%20.lookout.net/. Was http://www .lookout.net/.
FAIL canonicalize('http:// lookout.net/') should be http://%E1%9A%80lookout.net/. Was http:// lookout.net/.
FAIL canonicalize('http://lookout.net/') should be http://%1Flookout.net/. Was http://lookout.net/.
FAIL canonicalize('http://look۝out.net/') should be http://look%DB%9Dout.net/. Was http://look۝out.net/.
FAIL canonicalize('http://look᠎out.net/') should be http://look%E1%A0%8Eout.net/. Was http://look᠎out.net/.
FAIL canonicalize('http://look⁠out.net/') should be http://look%E2%81%A0out.net/. Was http://lookout.net/.
FAIL canonicalize('http://look﻿out.net/') should be http://look%EF%BB%BFout.net/. Was http://lookout.net/.
FAIL canonicalize('http://look🿾out.net/') should be http://look%F0%9F%BF%BEout.net/. Was http://look🿾out.net/.
FAIL canonicalize('http://look￺out.net/') should be http://look%EF%BF%BAout.net/. Was http://look￺out.net/.
FAIL canonicalize('http://look⿰out.net/') should be http://look%E2%BF%B0out.net/. Was http://look⿰out.net/.
FAIL canonicalize('http://looḱout.net/') should be http://look%CD%81out.net/. Was http://xn--looout-kp7b.net/.
FAIL canonicalize('http://look‮out.net/') should be http://look%E2%80%AEout.net/. Was http://look‮out.net/.
FAIL canonicalize('http://look⁫out.net/') should be http://look%E2%81%ABout.net/. Was http://look⁫out.net/.
FAIL canonicalize('http://look󠀁out.net/') should be http://look%F3%A0%80%81out.net/. Was http://look󠀁out.net/.
FAIL canonicalize('http://look󠀠out.net/') should be http://look%F3%A0%80%A0out.net/. Was http://look󠀠out.net/.
FAIL canonicalize('http://look־out.net/') should be http://look%D6%BEout.net/. Was http://look־out.net/.
PASS successfullyParsed is true

TEST COMPLETE
